Refactored API Tests for Details, Delete, Create comments, list comments
